fixing some errors and fix files

+++ ToffeeProject/src/Models/Candy.java
@@ -58,7 +58,7 @@ public class Candy{
         Name = name;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return Price;
     }
